chore: Enhance server1.nix with systemd-boot and sops-nix modules
Improve server1.nix by adding the systemd-boot bootloader and enabling automatic login for the user. This change enhances the boot process and user experience on the server. Additionally, include the sops-nix module to enable secure handling of secrets, enhancing the security of the application. Remove unnecessary comments and update the stateVersion.