chore: Add user account for 'arm' in server1.nix
Define a user account for 'arm' in the server1.nix file. This change includes the necessary configuration to create the 'arm' user, set it as a system user, and assign it to additional groups. The hashed password for the user is also provided. This addition enhances the server setup by including the 'arm' user account.